Custom Parser
Looks like we need to write a custom parser for this thing. We're indexing all this data as one flat document which is not taking advantage of the structured, hierarchical nature of EAD. Will tweak this parser to see if we can get something useful, adding more tags for more relevance as I test it.